requirements.txt
pandas>=1.5.0
numpy>=1.23.0,<2.0.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
xgboost>=1.7.0
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
pickle5>=0.0.11; python_version < '3.8'
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

